+++ src/main/java/chess/repository/dao/ChessGameDao.java
@@ -1,7 +1,8 @@
-package chess.dao;
+package chess.repository.dao;
 
 import chess.connection.DBConnectionUtil;
 import chess.domain.game.GameStatus;
+import chess.repository.ChessGameRepository;
 import chess.view.mapper.ColorMapper;
 
 import java.sql.Connection;
