+++ src/main/java/chess/service/ChessGameService.java
@@ -1,7 +1,7 @@
 package chess.service;
 
-import chess.dao.ChessGameRepository;
-import chess.dao.MovementRepository;
+import chess.repository.ChessGameRepository;
+import chess.repository.MovementRepository;
 import chess.domain.game.GameStatus;
 import chess.domain.square.Square;
 import chess.dto.Movement;
