+++ src/test/java/chess/dao/ChessGameDaoTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.domain.game.GameStatus;
 import chess.domain.piece.PieceColor;
+import chess.repository.ChessGameRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
